reorder classes (+barrel export), remove some unused imports
