+++ src/logic/tabs/index.ts
@@ -1,0 +1,12 @@
+export { Tab } from "./Tabs";
+export { CodeTab } from "./CodeTab";
+export { InheritanceViewTab } from "./InheritanceViewTab";
+
+export { openCodeTab } from "./Tabs";
+export { openInheritanceViewTab } from "./Tabs";
+export { openUnknownTypeTab } from "./Tabs";
+
+export { getOpenTab } from "./Tabs";
+export { closeTab } from "./Tabs";
+export { setTabPosition } from "./Tabs";
+export { closeOtherTabs } from "./Tabs";
